+++ packages/bochki_schedule_app/lib/src/domain/assistants/assistants_repository.dart
@@ -1,0 +1,6 @@
+import '../named_directory/named_directory_repository.dart';
+
+import 'assistant.dart';
+
+abstract interface class AssistantsRepository
+    implements NamedDirectoryRepository<Assistant> {}